+++ src/ErgoFab.Model/Organization.cs
@@ -27,7 +27,7 @@ namespace ErgoFab.Model
         public string Title { get; set; }
 
         [ForeignKey(nameof(TheCountry))]
-        public short? CountryId { get; set; }
+        public int? CountryId { get; set; }
         public Country TheCountry { get; set; }
